conversion of types numbers

+++ level10/lesson04/task03/Solution.java
@@ -13,7 +13,7 @@ public class Solution
     {
         float f = (float)128.50;
         int i = (int)f;
-        int b = (int)(i + f);
+        int b = (byte)(i + f);
         System.out.println(b);
     }
 }
